backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
cloudinary>=1.38.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Authentication
google-auth>=2.27.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
